chore: Update zsh and vim configurations
This commit updates the zsh and vim configurations. It sets the default editor to vim and adds an alias to use vim as vi. It also adds a new function `yy()` to the zshrc file, which creates a temporary file to store the current working directory and changes the directory to the one specified by the `yazi` command. This improves the functionality of the config files and enhances the user experience.